+++ hpx/runtime/threads/coroutines/coroutine.hpp
@@ -40,6 +40,7 @@
 #include <hpx/util/assert.hpp>
 
 #include <cstddef>
+#include <cstdint>
 #include <limits>
 #include <utility>
 
@@ -208,7 +209,7 @@ namespace hpx { namespace threads { namespace coroutines
 
         impl_ptr m_pimpl;
 
-        std::size_t count() const
+        std::uint64_t count() const
         {
             return m_pimpl->count();
         }
